refac:Process defined in mod.rs in a separate file
